introduced conjugate hierarchy
