+++ src/hierarchies/base_hierarchy.h
@@ -9,109 +9,10 @@
 #include <set>
 #include <stan/math/prim.hpp>
 
+#include "abstract_hierarchy.h"
 #include "hierarchy_id.pb.h"
 #include "marginal_state.pb.h"
 #include "src/utils/rng.h"
-
-//! Abstract base template class for a hierarchy object.
-
-//! This template class represents a hierarchy object in a generic iterative
-//! BNP algorithm, that is, a single set of unique values with their own prior
-//! distribution attached to it. These values are part of the Markov chain's
-//! state chain (which includes multiple hierarchies) and are simply referred
-//! to as the state of the hierarchy. This object also corresponds to a single
-//! cluster in the algorithm, in the sense that its state is the set of
-//! parameters for the distribution of the data points that belong to it. Since
-//! the prior distribution for the state is often the same across multiple
-//! different hierarchies, the hyperparameters object is accessed via a shared
-//! pointer. Lastly, any hierarchy that inherits from this class contains
-//! multiple ways of updating the state, either via prior or posterior
-//! distributions, and of evaluating the distribution of the data, either its
-//! likelihood (whose parameters are the state) or its marginal distribution.
-
-class AbstractHierarchy {
- protected:
-  std::set<int> cluster_data_idx;
-  int card = 0;
-  double log_card = stan::math::NEGATIVE_INFTY;
-
- public:
-  virtual ~AbstractHierarchy() = default;
-  virtual std::shared_ptr<AbstractHierarchy> clone() const = 0;
-  virtual bayesmix::HierarchyId get_id() const = 0;
-
-  //! Adds a datum and its index to the hierarchy
-  virtual void add_datum(
-      const int id, const Eigen::VectorXd &datum,
-      const bool update_params = false,
-      const Eigen::VectorXd &covariate = Eigen::VectorXd(0)) = 0;
-  //! Removes a datum and its index from the hierarchy
-  virtual void remove_datum(
-      const int id, const Eigen::VectorXd &datum,
-      const bool update_params = false,
-      const Eigen::VectorXd &covariate = Eigen::VectorXd(0)) = 0;
-  //! Deletes all data in the hierarchy
-  virtual void initialize() = 0;
-
-  virtual bool is_multivariate() const = 0;
-  virtual bool is_dependent() const { return false; }
-  virtual bool is_conjugate() const { return true; }
-  //!
-  virtual void update_hypers(
-      const std::vector<bayesmix::MarginalState::ClusterState> &states) = 0;
-
-  // EVALUATION FUNCTIONS FOR SINGLE POINTS
-  //! Evaluates the log-likelihood of data in a single point
-  virtual double like_lpdf(
-      const Eigen::RowVectorXd &datum,
-      const Eigen::RowVectorXd &covariate = Eigen::VectorXd(0)) = 0;
-
-  //! Evaluates the log-marginal distribution of data in a single point
-  virtual double prior_pred_lpdf(
-      const Eigen::RowVectorXd &datum,
-      const Eigen::RowVectorXd &covariate = Eigen::VectorXd(0)) = 0;
-
-  virtual double conditional_pred_lpdf(
-      const Eigen::RowVectorXd &datum,
-      const Eigen::RowVectorXd &covariate = Eigen::VectorXd(0)) = 0;
-
-  // EVALUATION FUNCTIONS FOR GRIDS OF POINTS
-  //! Evaluates the log-likelihood of data in a grid of points
-  virtual Eigen::VectorXd like_lpdf_grid(
-      const Eigen::MatrixXd &data,
-      const Eigen::MatrixXd &covariates = Eigen::MatrixXd(0, 0)) = 0;
-  //! Evaluates the log-marginal of data in a grid of points
-  virtual Eigen::VectorXd prior_pred_lpdf_grid(
-      const Eigen::MatrixXd &data,
-      const Eigen::MatrixXd &covariates = Eigen::MatrixXd(0, 0)) = 0;
-
-  virtual Eigen::VectorXd conditional_pred_lpdf_grid(
-      const Eigen::MatrixXd &data,
-      const Eigen::MatrixXd &covariates = Eigen::MatrixXd(0, 0)) = 0;
-
-  // SAMPLING FUNCTIONS
-  //! Generates new values for state from the centering prior distribution
-  virtual void sample_prior() = 0;
-  //! Generates new values for state from the centering posterior distribution
-  virtual void sample_full_cond(bool update_params = false) = 0;
-  virtual void write_state_to_proto(google::protobuf::Message *out) const = 0;
-  virtual void write_hypers_to_proto(google::protobuf::Message *out) const = 0;
-  virtual void set_state_from_proto(
-      const google::protobuf::Message &state_) = 0;
-
-  virtual void check_prior_is_set() = 0;
-
-  // GETTERS AND SETTERS
-  int get_card() const { return card; }
-  double get_log_card() const { return log_card; }
-  std::set<int> get_data_idx() { return cluster_data_idx; }
-  virtual google::protobuf::Message *get_mutable_prior() = 0;
-
-  //! Overloaded version of sample_full_cond(), mainly used for debugging
-  virtual void sample_full_cond(
-      const Eigen::MatrixXd &data,
-      const Eigen::MatrixXd &covariates = Eigen::MatrixXd(0, 0)) = 0;
-};
 
 template <class Derived, typename State, typename Hyperparams, typename Prior>
 class BaseHierarchy : public AbstractHierarchy {
@@ -122,7 +23,10 @@ class BaseHierarchy : public AbstractHierarchy {
   Hyperparams posterior_hypers;
   std::shared_ptr<Prior> prior;
 
-  virtual Hyperparams get_posterior_parameters() = 0;
+  std::set<int> cluster_data_idx;
+  int card = 0;
+  double log_card = stan::math::NEGATIVE_INFTY;
+
   virtual void initialize_state() = 0;
   void create_empty_prior() { prior.reset(new Prior); }
 
@@ -147,13 +51,6 @@ class BaseHierarchy : public AbstractHierarchy {
   void sample_prior() override {
     state = static_cast<Derived *>(this)->draw(*hypers);
   };
-  //! Generates new values for state from the centering posterior distribution
-  void sample_full_cond(bool update_params = true) {
-    Hyperparams params =
-        update_params ? static_cast<Derived *>(this)->get_posterior_parameters()
-                      : posterior_hypers;
-    state = static_cast<Derived *>(this)->draw(params);
-  }
 
   void initialize() override {
     hypers = std::make_shared<Hyperparams>();
@@ -162,10 +59,6 @@ class BaseHierarchy : public AbstractHierarchy {
     static_cast<Derived *>(this)->initialize_state();
     posterior_hypers = *hypers;
     static_cast<Derived *>(this)->clear_data();
-  }
-
-  void save_posterior_hypers() {
-    posterior_hypers = static_cast<Derived *>(this)->get_posterior_parameters();
   }
 
   void add_datum(
@@ -198,6 +91,10 @@ class BaseHierarchy : public AbstractHierarchy {
 
     return prior.get();
   }
+
+  int get_card() const override { return card; }
+  double get_log_card() const override { return log_card; }
+  std::set<int> get_data_idx() override { return cluster_data_idx; }
 
   virtual Eigen::VectorXd like_lpdf_grid(
       const Eigen::MatrixXd &data,
